delete client, edit ingredient, floats to long, client on input recalculations, input type number instead of text

+++ src/main/java/cz/client/dto/ClientDto.java
@@ -14,9 +14,9 @@ public class ClientDto implements Serializable {
     private Sex sex;
     private String firstName;
     private String secondName;
-    private int age;
-    private int weight;
-    private int height;
+    private long age;
+    private long weight;
+    private long height;
     private float bmrCoef;
     private BMRDto bmr;
